feat: helpPanel component lazy import로 분리
