+++ tsconfig.json
@@ -5,7 +5,8 @@
     "module": "esnext",
     "moduleResolution": "node",
     "lib": ["dom", "dom.iterable", "esnext"],
-    "noUnusedLocals": true,
+    "noUnusedLocals": false,
+    "noUnusedParameters": false,
     "allowJs": true,
     "sourceMap": true,
     "removeComments": true,
